requirements.txt
click
requests
python-dotenv
yaml
jenkins